# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.Actress:
    properties:
      aliases:
        description: Pipe-separated
        type: string
      created_at:
        type: string
      dmm_id:
        description: Real DMM actress ID when available (unique only for values >
          0)
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      translations:
        description: |-
          Translations carries optional translation data for API response enrichment.
          Not persisted in database — gorm:"-" prevents auto-migration/association.
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.ActressTranslation'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.ActressTranslation:
    properties:
      actress_id:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      language:
        type: string
      last_name:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api_actress.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  internal_api_actress.actressRequest:
    properties:
      aliases:
        type: string
      dmm_id:
        type: integer
      first_name:
        type: string
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
    type: object
...
  internal_api_actress.actressesResponse:
    properties:
      actresses:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.Actress'
        type: array
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_api_actress.bulkDeleteActressesRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  internal_api_actress.deleteActressesResponse:
    properties:
      deleted:
        type: integer
    type: object
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
  /api/v1/actresses:
    get:
      description: Get a paginated list of actresses with optional search query and
        sorting
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: name
        description: Sort column (id, dmm_id, japanese_name, first_name, last_name,
          created_at, updated_at, name)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction (asc, desc)
        in: query
        name: sort_order
        type: string
      - default: 50
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Skip results
        in: query
        name: offset
        type: integer
      - description: Language code to include translations for (e.g., 'en')
        in: query
        name: include_translations
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_actress.actressesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: List actresses
      tags:
      - actress
    post:
      consumes:
      - application/json
      description: Create a new actress record with the provided details
      parameters:
      - description: models.Actress details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_actress.actressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.Actress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Create actress
      tags:
      - actress
  /api/v1/actresses/{id}:
    delete:
      description: Delete an actress record by ID
      parameters:
      - description: models.Actress ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Delete actress
      tags:
      - actress
    get:
      description: Retrieve a single actress by their database ID
      parameters:
      - description: models.Actress ID
        in: path
        name: id
        required: true
        type: integer
      - description: Language code to include translations for (e.g., 'en')
        in: query
        name: include_translations
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.Actress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Get actress by ID
      tags:
      - actress
    put:
      consumes:
      - application/json
      description: Update an existing actress record by ID
      parameters:
      - description: models.Actress ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated actress details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_actress.actressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.Actress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Update actress
      tags:
      - actress
...
  /api/v1/actresses/bulk-delete:
    post:
      consumes:
      - application/json
      description: Delete multiple actress records (and their movie associations)
        by ID
      parameters:
      - description: Actress IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_actress.bulkDeleteActressesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_actress.deleteActressesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
      summary: Bulk delete actresses
      tags:
      - actress
  /api/v1/actresses/delete-all:
    post:
      description: Delete every actress record and all movie associations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_actress.deleteActressesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
      summary: Delete all actresses
      tags:
      - actress
...
schemes:
- http
- https
swagger: "2.0"